Fixup imports and split tests
