+++ flowmachine/flowmachine/core/server/server.py
@@ -3,7 +3,6 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import asyncio
-from concurrent.futures import Executor
 from json import JSONDecodeError
 import traceback
 
@@ -20,8 +19,7 @@ from marshmallow import ValidationError
 from zmq.asyncio import Context
 
 import flowmachine
-from flowmachine.core.cache import watch_and_shrink_cache
-from flowmachine.core.context import get_db, get_executor, action_request_context
+from flowmachine.core.context import action_request_context
 from flowmachine.utils import convert_dict_keys_to_strings
 from .exceptions import FlowmachineServerError
 from .zmq_helpers import ZMQReply
